+++ Sources/MapboxMaps/Style/Style.swift
@@ -25,8 +25,7 @@ public class Style {
      */
     public var uri: StyleURI = .streets {
         didSet {
-            let uriString = uri.rawValue.absoluteString
-            styleManager.setStyleURIForUri(uriString)
+            styleManager.setStyleURIForUri(uri.rawValue)
         }
     }